+++ hs-bindgen/app/HsBindgen/Cli/Info/UseDeclGraph.hs
@@ -49,7 +49,7 @@ parseOpts =
     Opts
       <$> parseConfig
       <*> parseUniqueId
-      <*> parseBaseModuleName
+      <*> parseBaseModuleName def
       <*> optional parseOutput'
       <*> parseInputs
       <*> parseFilePolicy