+++ hs-bindgen/app/HsBindgen/Cli/Internal/Frontend.hs
@@ -100,7 +100,7 @@ parseOpts =
       <$> parseDump
       <*> parseConfig
       <*> parseUniqueId
-      <*> parseBaseModuleName
+      <*> parseBaseModuleName def
       <*> parseInputs
       <*> parseFilePolicy